Restuctured web_api_testing directory
